Crate statemint_it

Re-exports

pub use xcm::v3::NetworkId::Kusama as KusamaId;

pub use xcm::v3::NetworkId::Polkadot as PolkadotId;

pub use xcm_emulator::cumulus_pallet_dmp_queue;

Modules

opaque

Macros

assert_expected_events

assert_ok

Panic if an expression doesn’t evaluate to Ok.

bx

Structs

AlwaysV2

WrapVersion implementation which attempts to always convert the XCM to version 2 before wrapping it.

AlwaysV3

WrapVersion implementation which attempts to always convert the XCM to version 3 before wrapping it.

Ancestor

A unit struct which can be converted into a MultiLocation of the inner parents value.

AncestorThen

A unit struct which can be converted into a MultiLocation of the inner parents value and the inner interior.

BHKusama

BHKusamaReceiver

BHKusamaSender

BHPolkadot

BHPolkadotReceiver

BHPolkadotSender

Collectives

CollectivesReceiver

CollectivesSender

InboundDownwardMessage

A wrapped version of DownwardMessage. The difference is that it has attached the block number when the message was sent.

Instance1

Instance1 to be used for instantiable palllets defined with the #[pallet] macro. Instances 2-16 are also available but are hidden from docs.

Kusama

KusamaMockNet

KusamaReceiver

KusamaSender

MultiAsset

Either an amount of a single fungible asset, or a single well-identified non-fungible asset.

MultiAssets

A Vec of MultiAssets.

MultiLocation

A relative path between state-bearing consensus systems.

PalletInfo

Parent

A unit struct which can be converted into a MultiLocation of parents value 1.

ParentThen

A tuple struct which can be converted into a MultiLocation of parents value 1 with the inner interior.

PenpalKusama

PenpalKusamaReceiver

PenpalKusamaSender

PenpalPolkadot

PenpalPolkadotReceiver

PenpalPolkadotSender

Polkadot

PolkadotMockNet

PolkadotReceiver

PolkadotSender

QueryResponseInfo

Information regarding the composition of a query response.

Statemine

StatemineReceiver

StatemineSender

Statemint

StatemintReceiver

StatemintSender

Weight

Xcm

XcmContext

Contextual data pertaining to a specific list of XCM instructions.

Enums

AssetId

Classification of an asset being concrete or abstract.

AssetInstance

A general identifier for an instance of a non-fungible asset class.

BodyId

An identifier of a pluralistic body.

BodyPart

A part of a pluralistic body.

Error

Error codes used in XCM. The first errors codes have explicit indices and are part of the XCM format. Those trailing are merely part of the XCM implementation; there is no expectation that they will retain the same index over time.

Fungibility

Classification of whether an asset is fungible or not, along with a mandatory amount or instance.

Instruction

Cross-Consensus Message: A message from one consensus system to another.

Junction

A single item in a path to describe the relative location of a consensus system.

Junctions

Non-parent junctions that can be constructed, up to the length of 8. This specific Junctions implementation uses a Rust enum in order to make pattern matching easier.

MaybeErrorCode

MultiAssetFilter

MultiAsset collection, defined either by a number of MultiAssets or a single wildcard.

NetworkId

A global identifier of a data structure existing within consensus.

OriginKind

Basically just the XCM (more general) version of ParachainDispatchOrigin.

Outcome

Outcome of an XCM execution.

Response

Response data to a query.

SendError

Error result value when attempting to send an XCM message.

Unsupported

VersionedAssetId

A single version’s Response value, together with its version code.

VersionedInteriorMultiLocation

A single InteriorMultiLocation value, together with its version code.

VersionedMultiAsset

A single MultiAsset value, together with its version code.

VersionedMultiAssets

A single MultiAssets value, together with its version code.

VersionedMultiLocation

A single MultiLocation value, together with its version code.

VersionedResponse

A single version’s Response value, together with its version code.

VersionedXcm

A single XCM message, together with its version code.

WeightLimit

An optional weight limit.

WildFungibility

Classification of whether an asset is fungible or not.

WildMultiAsset

A wildcard representing a set of assets.

XcmError

Error codes used in XCM. The first errors codes have explicit indices and are part of the XCM format. Those trailing are merely part of the XCM implementation; there is no expectation that they will retain the same index over time.

Constants

ALICE

BOB

POLKADOT_ED

PROOF_SIZE_THRESHOLD

REF_TIME_THRESHOLD

XCM_V3

XCM_VERSION

This module’s XCM version.

Traits

BHKusamaPallet

BHPolkadotPallet

CollectivesPallet

Encode

Trait that allows zero-copy write of value-references to slices in LE format.

ExecuteXcm

Type of XCM message executor.

Inspect

Trait for providing balance-inspection access to a set of named fungible assets.

IntoVersion

Attempt to convert self into a particular version of itself.

KusamaPallet

Para

PolkadotPallet

PreparedMessage

Relay

SendXcm

Utility for sending an XCM message to a given location.

StateminePallet

StatemintPallet

TestExt

Unwrappable

WrapVersion

Convert an Xcm datum into a VersionedXcm, based on a destination MultiLocation which will interpret it.

XcmWeightInfo

Functions

send_xcm

Convenience function for using a SendXcm implementation. Just interprets the dest and wraps both in Some before passing them as as mutable references into T::send_xcm.

validate_send

Convenience function for using a SendXcm implementation. Just interprets the dest and wraps both in Some before passing them as as mutable references into T::send_xcm.

weight_within_threshold

Type Definitions

AccountId

Some way of identifying an account on the chain. We intentionally make it equivalent to the public key of our transaction signing scheme.

AlwaysLatest

WrapVersion implementation which attempts to always convert the XCM to the latest version before wrapping it.

AlwaysLts

WrapVersion implementation which attempts to always convert the XCM to the most recent Long- Term-Support version before wrapping it.

InteriorMultiLocation

A relative location which is constrained to be an interior location of the context.

QueryId

An identifier for a query.

SendResult

Result value when attempting to send an XCM message.

XcmHash

A hash type for identifying messages.

XcmResult

XcmVersion

A version of XCM.

Derive Macros

Encode

Derive parity_scale_codec::Encode and parity_scale_codec::EncodeLike for struct and enum.